+++ buildSrc/src/main/kotlin/Configurations.kt
@@ -38,10 +38,10 @@ object Versions {
 
     object Sdk {
         const val classpath = "4.0.0"
-        const val buildTools = "29.0.3"
-        const val compile = 29
+        const val buildTools = "30.0.1"
+        const val compile = 30
         const val min = 21
-        const val target = 29
+        const val target = 30
     }
 
     val java = JavaVersion.VERSION_1_8
@@ -78,11 +78,11 @@ object Versions {
 
     object Firebase {
         const val classpath = "4.3.3"
-        const val analytics = "17.4.3"
+        const val analytics = "17.4.4"
 
         object Crashlytics {
             const val classpath = "2.2.0"
-            const val crashlytics = "17.1.0"
+            const val crashlytics = "17.1.1"
         }
     }